--- src/components/Animation/Animation.components.css
@@ -1,42 +1,0 @@
-.container {
-    width: 100%;
-    height: 100%;
-    position: relative;
-    display: block;
-    margin: auto
-}
-
-.image {
-    position: absolute;
-    top: 0;
-    left: 0;
-    z-index: 0;
-}
-
-.hovertext {
-    position: absolute;
-    top: auto;
-    left: 5%;
-    z-index: 1;
-    opacity: 0;
-}
-
-.container:hover .hovertext {
-    animation: showFromDown 0.5s ease-out forwards;
-}
-
-@keyframes showFromDown {
-    0% {
-        top: 100px;
-        opacity: 0;
-    }
-
-    75% {
-        opacity: 1;
-    }
-
-    100% {
-        top: 50px;
-        opacity: 1;
-    }
-}
