+++ src/components/index.js
@@ -1,2 +1,1 @@
 export * from './Button/Button.components'
-export * from './Animation/Animation.components'
