--- src/figmalocify/AndroidSmall1.css
@@ -1,126 +1,0 @@
-@keyframes scale-up {
-  0% {
-    transform: scale(0.5);
-  }
-  to {
-    transform: scale(1);
-  }
-}
-.a-shoe-brand {
-  margin: 0;
-  position: absolute;
-  top: 269px;
-  left: 88px;
-  display: inline-block;
-  width: 623px;
-  height: 126px;
-}
-.image-3-icon {
-  position: absolute;
-  top: calc(50% - 212px);
-  left: calc(50% - 230px);
-  width: 460px;
-  height: 256px;
-  object-fit: cover;
-  opacity: 1;
-}
-.image-3-icon.animate {
-  animation: 1s ease 0s 1 normal forwards scale-up;
-}
-.a-shoe-brand-which-is-a-world-parent {
-  position: absolute;
-  width: 100%;
-  top: calc(50% + 50px);
-  right: 0;
-  left: 0;
-  height: 450px;
-  overflow: hidden;
-}
-.android-small-1-child {
-  position: absolute;
-  width: 100%;
-  top: 0;
-  right: 0;
-  left: 0;
-  max-width: 100%;
-  overflow: hidden;
-  height: 570px;
-  object-fit: cover;
-}
-.android-small-1 {
-  position: relative;
-  background-color: #f18585;
-  width: 100%;
-  height: 1000px;
-  overflow: hidden;
-  cursor: auto;
-  text-align: left;
-  font-size: 40px;
-  color: #000;
-  font-family: var(--font-inter);
-}
-@media screen and (max-width: 1200px) {
-  .a-shoe-brand {
-    font-size: 24px;
-    width: 750px;
-    height: 130px;
-  }
-  .image-3-icon {
-    display: flex;
-    width: 350px;
-    height: 200px;
-    cursor: ew-resize;
-  }
-  .android-small-1-child {
-    width: 1000px;
-    height: 600px;
-  }
-  .android-small-1 {
-    width: 1000px;
-    height: 1200px;
-    cursor: auto;
-  }
-}
-@media screen and (max-width: 960px) {
-  .a-shoe-brand {
-    font-size: 30px;
-    width: 700px;
-    height: 100px;
-  }
-  .image-3-icon {
-    width: 275px;
-    height: 150px;
-    cursor: ew-resize;
-  }
-  .android-small-1-child {
-    width: 900px;
-    height: 525px;
-  }
-  .android-small-1 {
-    width: 900px;
-    height: 1050px;
-  }
-}
-@media screen and (max-width: 420px) {
-  .a-shoe-brand {
-    font-size: 20px;
-    width: 300px;
-    height: 75px;
-  }
-  .image-3-icon {
-    width: 200px;
-    height: 100px;
-    display: block;
-    align-items: center;
-    justify-content: center;
-    margin: auto;
-  }
-  .android-small-1-child {
-    width: 400px;
-    height: 300px;
-  }
-  .android-small-1 {
-    width: 400px;
-    height: 800px;
-  }
-}
