--- src/figmalocofy2/Module1.css
@@ -1,215 +1,0 @@
-@keyframes slide-in-top {
-  0% {
-    transform: translateY(-200px);
-    opacity: 0;
-  }
-  to {
-    transform: translateY(0);
-    opacity: 1;
-  }
-}
-.fashion-meets-comfort {
-  display: block;
-}
-.a-chic-addition-to-your-look {
-  display: block;
-  font-size: var(--font-size-base);
-}
-.description-txt {
-  line-break: anywhere;
-  width: 100%;
-}
-.description,
-.module1left {
-  width: 700px;
-  height: 500px;
-  flex-shrink: 0;
-  display: flex;
-  align-items: center;
-}
-.description {
-  margin: 0;
-  position: relative;
-  font-weight: 600;
-}
-.module1left {
-  overflow: hidden;
-  flex-direction: column;
-  justify-content: flex-end;
-  scale: 0.75;
-}
-.hover-description,
-.slipper-img-icon {
-  position: absolute;
-  left: calc(50% - 350px);
-  width: 700px;
-  height: 500px;
-}
-.slipper-img-icon {
-  top: calc(50% - 250px);
-  object-fit: cover;
-  scale: 1;
-}
-.hover-description {
-  margin: 0;
-  top: calc(50% - 249px);
-  font-weight: 600;
-  display: flex;
-  align-items: center;
-  text-shadow: 0 4px 4px rgba(0, 0, 0, 0.25);
-  opacity: 0;
-  background-color: var(--color-darkgray);
-  padding: var(--padding-md);
-}
-.hover-description:hover {
-  display: flex;
-}
-.module1,
-.module1right {
-  position: relative;
-  overflow: hidden;
-}
-.module1right {
-  width: 700px;
-  height: 500px;
-  flex-shrink: 0;
-  text-align: center;
-  font-size: 60%;
-  color: var(--color-gray);
-}
-.module1right:hover .hover-description {
-  animation: 1s ease 0s 1 alternate backwards slide-in-top;
-  opacity: 1;
-  scale: 1;
-}
-.module1right:hover .slipper-img-icon {
-  opacity: 0.5;
-}
-.module1 {
-  background-color: var(--color-lightcoral);
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  padding: 0 0 0 5px;
-  box-sizing: border-box;
-  align-items: center;
-  justify-content: center;
-  text-align: left;
-  font-size: 30px;
-  color: var(--color-white);
-  font-family: var(--font-inter);
-}
-@media screen and (max-width: 1200px) {
-  .description {
-    text-align: left;
-  }
-  .module1left {
-    flex-direction: row;
-    gap: var(--gap-0);
-    align-items: center;
-    justify-content: flex-end;
-  }
-  .slipper-img-icon {
-    animation: show-on-hover 1s ease infinite forwards;
-  }
-  .hover-description {
-    display: flex;
-    opacity: 0;
-  }
-  .hover-description:hover {
-    opacity: 1;
-  }
-  .module1 {
-    width: auto;
-    height: auto;
-    flex-direction: row;
-  }
-}
-@media screen and (max-width: 960px) {
-  .module1left {
-    flex-direction: column;
-  }
-  .module1 {
-    flex-direction: row;
-  }
-}
-@media screen and (max-width: 420px) {
-  .module1left {
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 0;
-    box-sizing: border-box;
-    scale: 0.55;
-  }
-  .slipper-img-icon {
-    scale: 0.6;
-  }
-  .hover-description {
-    opacity: 0;
-    font-size: 75%;
-  }
-  .module1right {
-    height: 300px;
-  }
-  .module1 {
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-  }
-}
-@media screen and (max-width: 599px) {
-  .module1left {
-    flex-direction: row;
-    padding: 0;
-    box-sizing: border-box;
-  }
-  .module1 {
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-}
-@media screen and (min-width: 600px) {
-  .module1left {
-    flex-direction: row;
-    gap: var(--gap-0);
-    padding: 0;
-    box-sizing: border-box;
-  }
-  .module1 {
-    flex-direction: column;
-  }
-}
-@media screen and (min-width: 768px) {
-  .module1left {
-    gap: var(--gap-0);
-    padding: 0;
-    box-sizing: border-box;
-  }
-  .module1 {
-    flex-direction: column;
-  }
-}
-@media screen and (min-width: 992px) {
-  .module1left {
-    gap: var(--gap-0);
-    padding: 0;
-    box-sizing: border-box;
-  }
-  .module1,
-  .module1left {
-    flex-direction: row;
-  }
-}
-@media screen and (min-width: 1200px) {
-  .module1left {
-    gap: var(--gap-0);
-    padding: 0;
-    box-sizing: border-box;
-  }
-  .module1 {
-    width: 1200px;
-    flex-direction: row;
-  }
-}
