--- src/iframe_components/Text.components.css
@@ -1,7 +1,0 @@
-.text {
-    color: red;
-}
-
-.text:hover {
-    color: blue;
-}
