--- src/shadowroot_components/JSSelector/Anim.components.js
@@ -1,50 +1,0 @@
-import { useEffect, useRef } from 'react';
-import { ShadowRoot } from '../../wrappers/ShadowRoot';
-import './Anim.css'
-
-function Util(props) {
-
-    const elements = useRef(null);
-
-    useEffect(() => {
-        console.log(elements)
-        elements.current.addEventListener('mouseover', () => {
-            elements.current.classList.add('change-color')
-        });
-        elements.current.addEventListener('mouseout', () => {
-            elements.current.classList.remove('change-color')
-        });
-    }, []);
-
-
-    return (
-        <>
-            <style>
-                {require('!raw-loader!./Anim.css').default}
-            </style>
-            <button className="emoji container" ref={elements}>
-                <h2>
-                    Im inside shadow root and i am unable to be animated with js selectors, so I am now animated with React.useRef. if you move mouse out of me I will be hidden
-                </h2>
-            </button>
-        </>
-    )
-
-}
-
-
-JSAnim.defaultProps = {
-
-};
-
-
-export default function JSAnim() {
-    return (
-        <>
-            Im not in shadow dom
-            <ShadowRoot>
-                <Util />
-            </ShadowRoot>
-        </>
-    )
-}
