--- src/shadowroot_components/JSSelector/Anim.css
@@ -1,25 +1,0 @@
-.container {
-    background-color: white;
-    color: white;
-    width: 50vw;
-}
-
-.change-color h2 {
-    position: relative;
-    animation: slide-from-right 0.6s 1 ease-in-out forwards;
-    background-color: white;
-    color: green;
-    overflow: hidden;
-}
-
-@keyframes slide-from-right {
-    0% {
-        opacity: 0;
-        transform: translateX(400px);
-    }
-
-    100% {
-        opacity: 1;
-        transform: translateX(0px);
-    }
-}
