+++ src/stories/Button.stories.js
@@ -1,25 +1,25 @@
-import React from 'react';
-import Button from '../components/Button/Button.components';
+// import React from 'react';
+// import Button from '../components/Button/Button.components';
 
-export default {
-  title: 'Components/Buttons',
-  component: Button,
-  argTypes: {
-    backgroundColor: { control: 'color' },
-  },
-};
+// export default {
+//   title: 'Components/Buttons',
+//   component: Button,
+//   argTypes: {
+//     backgroundColor: { control: 'color' },
+//   },
+// };
 
-const Template = (args) => <Button {...args} />;
+// const Template = (args) => <Button {...args} />;
 
-export const Primary = Template.bind({});
-Primary.args = {
-  primary: true,
-  label: 'Button',
-};
+// export const Primary = Template.bind({});
+// Primary.args = {
+//   primary: true,
+//   label: 'Button',
+// };
 
-export const Secondary = Template.bind({});
-Secondary.args = {
-  label: 'Button',
-  backgroundColor: 'grey',
-};
+// export const Secondary = Template.bind({});
+// Secondary.args = {
+//   label: 'Button',
+//   backgroundColor: 'grey',
+// };
