--- src/stories/ShadowAnim.stories.js
@@ -1,11 +1,0 @@
-import JSAnim from "../shadowroot_components/JSSelector/Anim.components";
-
-export default {
-    title: 'Components/Shadow Root/JS Selector/Anim',
-    component: JSAnim
-}
-
-const Template = (args) => <JSAnim {...args} />;
-
-export const Default = Template.bind({});
-Default.args = {};
