+++ src/stories/ZiglyM1.stories.js
@@ -1,0 +1,10 @@
+import ZiglyM1 from '../components/zigly/Module1'
+
+export default {
+    title: 'Components/Zigly/Module 1',
+    component: ZiglyM1,
+}
+
+const Template = (args) => <ZiglyM1 {...args} />
+
+export const Primary = Template.bind({})
