--- src/stories/figmalocofy.stories.js
@@ -1,14 +1,0 @@
-import React from "react";
-import AndroidSmall1 from "../figmalocify/AndroidSmall1";
-
-export default {
-    title: 'Components/FigmaLocofy1',
-    component: AndroidSmall1
-}
-
-const Template = (args) => <AndroidSmall1 {...args} />;
-
-
-export const Primary = Template.bind({});
-Primary.args = {
-};
