--- src/stories/figmalocofy2.stories.js
@@ -1,14 +1,0 @@
-import React from "react";
-import Module1 from "../figmalocofy2/Module1";
-
-export default {
-    title: 'Components/FigmaLocofy2/Cai Store',
-    component: Module1
-}
-
-const Template = (args) => <Module1 {...args} />;
-
-
-export const Primary = Template.bind({});
-Primary.args = {
-};
